Refactor and move around
